requirements.txt
django
line-bot-sdk
requests
djangorestframework
django-filter  
markdown
